% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCPO-package.R
\docType{package}
\name{DCPO-package}
\alias{DCPO-package}
\alias{DCPO}
\title{DCPO: Dynamic Comparative Public Opinion}
\description{
DCPO estimates dynamic comparative public opinion as a latent variable from survey data
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
