% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpo.R
\name{dcpo}
\alias{dcpo}
\title{Estimate Dynamic Comparative Public Opinion}
\usage{
dcpo(x, chime = TRUE, ...)
}
\arguments{
\item{x}{a data frame of survey items and marginals generated by \code{DCPOtools::dcpo_setup}}

\item{chime}{play chime when complete?}
}
\value{
a stanfit object
}
\description{
\code{dcpo} uses diverse survey data to estimate public opinion across countries and over time.
}
\details{
\code{dcpo}, when passed a data frame \code{x} of survey marginals created by \code{dcpo_setup}, estimates a latent variable of public opinion.
}
