% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_surveys.R
\name{get_surveys}
\alias{get_surveys}
\title{Get Survey Datasets for DCPO}
\usage{
get_surveys(vars, datapath = "../data/dcpo_surveys", file = "",
  chime = TRUE)
}
\arguments{
\item{vars}{a data frame (or, optionally, a vector of survey names or a .csv file) of survey items}

\item{datapath}{path to the directory that will house the raw survey datasets}

\item{chime}{play chime when complete?}
}
\value{
the function downloads datasets
}
\description{
\code{get_surveys} downloads the survey datasets to be used to estimate public opinion across countries and over time
}
\details{
\code{get_surveys}, when passed a data frame of survey items, downloads the
source survey datasets, converts them to .RData format, and saves them to a specified
directory for later use by \code{dcpo_setup}.  When constructing a list of surveys, one
should be use the name employed in \code{DCPO}'s built-in \code{surveys_data} database:
\code{View(surveys_data)}.
}
