% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys_data.R
\docType{data}
\name{surveys_data}
\alias{surveys_data}
\title{Surveys Used As Sources for DCPO}
\format{A dataframe with six columns}
\usage{
surveys_data
}
\description{
A dataframe of surveys used internally by \code{dcpo_setup}
that contains the following variables:
}
\details{
\itemize{
  \item survey
  \item load_cmd
  \item cy_data
  \item wt
  \item notes
  \item source
}
}
\keyword{datasets}

