% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_min_yrs.R
\name{with_min_yrs}
\alias{with_min_yrs}
\title{Select Countries with Observations in a Minimum Number of Distinct Years}
\usage{
with_min_yrs(x, min_yrs)
}
\arguments{
\item{x}{a data frame of survey items and marginals generated by \code{dcpo_setup}}

\item{min_yrs}{the minimum number of distinct years needed for a country included in x to be retained}
}
\value{
a data frame
}
\description{
\code{with_min_yrs} is a convenience function for filtering data to be input to \code{dcpo} by the number of country-years observed
}
