% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpo.R
\name{dcpo}
\alias{dcpo}
\title{Estimate Dynamic Comparative Public Opinion}
\usage{
dcpo(x, scale_item, chime = TRUE, ...)
}
\arguments{
\item{x}{a data frame of survey items and marginals generated by \code{dcpo_setup}}

\item{scale_item}{a string specifying a question-cutpoint to identify the scale of the public opinion; this item will be assigned a difficulty of .5}

\item{chime}{play chime when complete?}

\item{seed}{an integer specifying the seed for random number generation for replication purposes}

\item{iter}{a positive integer specifying the number of iterations for each chain (including warmup)}

\item{chains}{a positive integer specifying the number of Markov chains; the default is 3}

\item{cores}{a positive integer specifying the number of processor cores to use; the default is the lesser of the number of chains or half the available cores}

\item{adapt_delta}{a double between 0 and 1; the default is .95.  Lower values complete faster at a greater risk of nonconvergence}
}
\value{
a stanfit object
}
\description{
\code{dcpo} uses diverse survey data to estimate public opinion across countries and over time.
}
\details{
\code{dcpo}, when passed a data frame \code{x} of survey marginals created by \code{dcpo_setup}, estimates a latent variable of public opinion.

If a data frame of independent variables including observations for all countries that appear in \code{x} for each year in the time period from the first to last year to appear in \code{x} is passed to the argument \code{iv}, \code{dcpo} will also estimate coefficients for these predictors.
}
