% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpo.R
\name{dcpo}
\alias{dcpo}
\title{Estimate Dynamic Comparative Public Opinion}
\usage{
dcpo(x, seed = 324, iter = 3000, cores = min(4,
  parallel::detectCores()/2), chains = 4, adapt_delta = 0.95,
  robust = FALSE, constant_alpha = FALSE, chime = TRUE)
}
\arguments{
\item{df}{a data frame of survey items and marginals generated by \code{dcpo_setup}}
}
\value{
a stanfit object
}
\description{
\code{dcpo} uses diverse survey data to estimate public opinion across countries and over time.
}
\details{
\code{dcpo_setup}, when passed a data frame of survey items, collects the
responses and formats them for use with the \code{dcpo} function.
}
