% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpo_setup.R
\name{dcpo_setup}
\alias{dcpo_setup}
\title{Prepare Data for DCPO}
\usage{
dcpo_setup(vars, datapath = "../data/dcpo_surveys", file = "",
  chime = TRUE)
}
\arguments{
\item{vars}{a data frame (or, optionally, a .csv file) of survey items}

\item{datapath}{path to the directory that houses raw survey datasets}

\item{file}{a file path to save output to (in comma-separated format)}

\item{chime}{play chime when complete?}
}
\value{
a data frame
}
\description{
\code{dcpo_setup} prepares survey data for use with the \code{dcpo} function.
}
\details{
\code{dcpo_setup}, when passed a data frame of survey items, collects the
responses and formats them for use with the \code{dcpo} function.
}
