% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{dcpo_setup}
\alias{dcpo_setup}
\title{Prepare Data for DCPO}
\usage{
dcpo_setup(vars, datapath = "~/Documents/Projects/Data/")
}
\arguments{
\item{vars}{a data frame of survey items}

\item{datapath}{path to the directory that houses raw survey datasets}
}
\value{
a data frame
}
\description{
\code{dcpo_setup} prepares survey data for use with the \code{dcpo} function.
}
\details{
\code{dcpo_setup}, when passed a data frame of survey items, collects the
responses and formats them for use with the \code{dcpo} function.
}

