% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcpo_xvt.R
\name{dcpo_xvt}
\alias{dcpo_xvt}
\title{Cross-validation testing for DCPO}
\usage{
dcpo_xvt(
  dcpo_input,
  fold_number = 1,
  number_of_folds = 10,
  fold_seed = 324,
  chime = TRUE,
  ...
)
}
\arguments{
\item{dcpo_input}{a data frame of survey items and marginals generated by \code{DCPOtools::dcpo_setup}}

\item{fold_number}{an integer indicating the number of the fold to treated as test data in the current analysis}

\item{number_of_folds}{an integer indicating the total number of folds}

\item{fold_seed}{a seed for reproducibly randomly assigning observations to folds; when a complete set of k-fold cross-validations is to be performed, the same seed should be used for all}

\item{chime}{play chime when complete?}

\item{...}{arguments to be passed to \code{rstan::stan}. See \code{dcpo}.}
}
\value{
a stanfit object
}
\description{
\code{dcpo_xvt} performs a single cross-validation test for DCPO
}
\details{
\code{dcpo_xvt} performs a single cross-validation test of a DCPO estimation.  To perform
a complete k-fold cross-validation, call it repeatedly, changing only the fold_number argument.
}
\examples{
\donttest{
# Single cross-validation test with 25% test set
demsup_xvtest_25pct <- dcpo_xvt(demsup_data,
                           number_of_folds = 4,
                           iter = 300,
                           chains = 1) # 1 chain/300 iterations for example only; use defaults
}

}
