% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_min_yrs.R
\name{with_min_yrs}
\alias{with_min_yrs}
\title{Select Countries with Observations in a Minimum Number of Distinct Years}
\usage{
with_min_yrs(x, min_yrs)
}
\arguments{
\item{x}{a data frame of survey items and marginals generated by \code{dcpo_setup}}

\item{min_yrs}{the minimum number of distinct years needed for a country included x to be retained}
}
\value{
a data frame
}
\description{
\code{dcpo} uses diverse survey data to estimate public opinion across countries and over time.
}
\details{
\code{dcpo}, when passed a data frame \code{x} of survey marginals created by \code{dcpo_setup}, estimates a latent variable of public opinion.

If a data frame of independent variables including observations for all countries that appear in \code{x} for each year in the time period from the first to last year to appear in \code{x} is passed to the argument \code{iv}, \code{dcpo} will also estimate coefficients for these predictors.
}
